Created new simplified build system. The firmware makefile is now the master of the firmware build process, with the cmake configuration and building of the shared library hidden away from the user. CMake remains the way to configure the GUI build.
Split up the avionics_config into smaller files

+++ Avionics.GUI/Source/SerialPort.cpp
@@ -56,13 +56,13 @@ int SerialPort::read(uint8_t* buff, size_t size) {
 
 	if (size > 4096)
 		size = 4096;
-	ReadFile(port_handle_, buff, size, &n, nullptr);
+	ReadFile(port_handle_, buff, (DWORD)size, &n, nullptr);
 	return n;
 }
 
 int SerialPort::write(const uint8_t* buff, size_t size) {
 	FTAssert(portIsOpen(), "Trying to write to closed port");
 	DWORD n;
-	WriteFile(port_handle_, buff, size, &n, nullptr);
+	WriteFile(port_handle_, buff, (DWORD)size, &n, nullptr);
 	return (int)n;
 }
